SP5 IAxis: pixel units should be float

+++ src/ScottPlot5/ScottPlot5/AxisPanels/XAxisBase.cs
@@ -103,7 +103,7 @@ public abstract class XAxisBase : AxisBase, IAxis
         return distance * dataArea.Width / Width;
     }
 
-    public double GetCoordinateDistance(double distance, PixelRect dataArea)
+    public double GetCoordinateDistance(float distance, PixelRect dataArea)
     {
         return distance / (dataArea.Width / Width);
     }
